cloud_run_app/requirements.txt
flask
tensorflow
numpy
google-cloud-storage
gunicorn  # Servidor WSGI recomendado para producción en Cloud Run